Android Part 2, Home Work, Lesson 7

+++ app/src/main/java/ru/geekbrains/kozirfm/weatherapp/data/Wind.java
@@ -7,7 +7,7 @@ public class Wind {
 
     @SerializedName("speed")
     @Expose
-    private int speed;
+    private float speed;
     @SerializedName("deg")
     @Expose
     private int deg;
@@ -20,7 +20,7 @@ public class Wind {
         this.deg = deg;
     }
 
-    public int getSpeed() {
+    public float getSpeed() {
         return speed;
     }
